+++ CHANGES.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* New islands validation rule: No pieces on invisible hexagons
+
 ### Changed
 
 * Make it easier to read the upkeep cost in the tutorial screen by adding a white border to the numbers